Completed main module set up and added "id" field to config.json
Completed set up of main module to only send emails once a day if slots are found. Added done set, and set up a reset time for the done set to be reset. Added "id" field in the config file and to the config reader module. Made the "id" field mandatory